Initial Filelist Support
Adds support for supplying a file with a list of files in it. Useful for getting around 8k cmd character limit.

Other options would be to reparse a existing CIFINISH to append instead of replacing on rerun, or allowing to supply a directory for cia var.
This is my first python work, I do batch at most... so this is my best.